Moved some helper functions from version.helper to release.helper and some functions from version.version to version.helper
